requirements.txt
wheel
dialog-bot-sdk
grpcio
python-dotenv